requirements/base.txt
PyYAML==6.0
pydantic[dotenv]==1.10.4
fastapi==0.100.0
uvicorn[standard]==0.22.0
tinydb==4.8.0
google-api-python-client==2.92.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
